InventoryEntryDraft to interface. #942
